Pagecache nodes: replace dirty_commits queue with operation list
Using a limited-size queue to store pagecache node sync requests
can lead to a failure to enqueue a request if the user program
issues many parallel fsync() syscalls on a given file. In addition,
using this queue to store both commit closures and sync completion
status handlers can cause failure to clear the `committing` flag in
the pagecache_node struct (because a completion status handler will
not call the commit_dirty_node_complete() function, which is where
the flag is cleared), which can lead to a deadlock where fsync()
calls by the user program will never return.
This change replaces the limited-size queue and the `committing`
flag with an operation list where commit closures and completion
status handlers are stored (wrapped in a struct which contains
information on the type of operation); a commit operation is only
removed from the list when the operation is completed.

Closes #1937