Unix clocks: add support for {clock,pthread}_getcpuclockid()
The clock_gettime(), clock_getres() and timer_create() syscalls can
be invoked with clock id values returned by the
clock_getcpuclockid() and pthread_getcpuclockid() C library
functions.
This change modifies the logic of the above syscalls so that they
can properly decode such clock id values. For POSIX timers, in case
a timer is created with a clock id corresponding to the CPU time of
an arbitrary thread, a reference to this thread is kept in the
timer struct, because the thread's timer queue must not be
deallocated until the timer is closed.

Closes #1946