vm_exit(): ensure vm_shutdown() is not called after vm_halt
Under Qemu, the ACPI halt handler does not halt the VM immediately,
thus the guest code may continue to execute after the halt request;
this may cause the vm_shutdown() to be called after the halt
handler, with the effect that Qemu exits with a non-zero code (on
x86_64) even if the "debug_exit" option is not present in the
manifest.
This change adds a machine_halt() call after the invocation of the
halt handler, so that no other code is executed after the halt
handler (the infinite loop in the Hyper-V halt handler is being
removed since it's no longer necessary). This fixes spurious "exit
status 1" error messages output by Ops when running on-prem
instances.