Aarch64: add halt_handler closure implementation
In an Aarch64 VM, if the Angel shutdown command is not supported,
it generates an exception, which is logged by the kernel. With the
existing code, when shutting down a VM which does not support the
Angel shutdown command (for example, on Apple M1 and M2 machines),
a "QEMU exit code will not reflect program exit code" error message
is output by the kernel, even if the "debug_exit" option is not
specified in the manifest.
This change adds a halt_handler implementation to the Aarch64
platform, so that the Angel shutdown command is only executed if
the "debug_exit" option is specified. This removes the above error
message when the option is not specified. In addition, the
`-t "(debug_exit:t)"` command line option for the mkfs invocation
in the Makefile is being moved before the `$(IMAGE)` argument: this
allows this option to be properly parsed on OSes (such as MacOS)
where option arguments specified after non-option arguments are
ignored.