mcache: make "malloc-style" allocation support optional
In the mcache implementaion, the meta heap is used to operate the
fallback table, which is needed to keep track of the size of
allocation requests when deallocations have an unknown size.
Among the kernel heaps, the general heap is initialized to use the
bootstrap heap as meta heap; use of the bootstrap heap by the
general heap after initialization creates concurrency problems,
because the bootstrap heap is also used by the physical heap.
This change adds a "malloc_style" boolean parameter to the
allocate_mcache() function, so that an mcache can be configured to
not use the malloc-style code (and thus to not use the meta heap
after initialization). The only heap in the kernel that needs the
malloc-style functionality is the "malloc" heap, which is now
declared among the other kernel heaps and used by vendor code
(lwIP, mbedTLS, ACPI).