x86 MMU: allow init_mmu() to be called by UEFI loader
This change makes the init_mmu(void) function available to be
called by the UEFI loader, and refactors it to check for the
maximum supported cpuid function and to make use of symbolic
constants. This function is now called by both the UEFI loader and
the kernel.
This change will be needed when adding support for memory
encryption.