linear backed heap: add support for pre-mapped virtual memory range
With this change, it is possible to instantiate a linear backed
heap that uses a generic, pre-mapped virtual address space instead
of the hard-coded address space defined by the LINEAR_BACKED_BASE
and LINEAR_BACKED_PHYSLIMIT constants. This will be needed to
create a separate linear heap to be used for memory encryption.