vsock_events(): do not acquire vsock_connection lock
Currently, the events closure of vsock file descriptors is locking
the underlying vsock_connection in order to avoid calling the
virtio_sock_get_buf_space() function in the middle of an update to
the connection counters from which the buffer space is calculated.
However, the events closure is called from the Unix epoll code
with the epollfd lock held; this creates a lock inversion issue if
the events closure is invoked at the same time as another CPU calls
notify_dispatch() for a vsock file descriptor (with the
vsock_connection locked) registered in an epoll instance.
This change fixes the lock inversion issue by not locking the
vsock_connection in the events closure: this is safe because the
counters (from which the buffer space is calculated) that are
updated during a connection are tx_cnt and peer_fwd_cnt, and tx_cnt
is always greater than or equal to peer_fwd_cnt (unless the virtio
socket device is misbehaving by reporting an invalid fwd_cnt
value), thus virtio_sock_get_buf_space() always returns a
consistent value even without locking.