Virtio socket: ensure received messages are processed in order
In order to satisfy the requirements of the stream socket type,
which guarantees delivery of ordered packets during a connection,
the virtio socket code must process all messages received from the
device in the order they are received. With the existing code, in
SMP machines it is possible for a virtio_sock_rxbuf completion to
be invoked before the invocation of the completion of a previously
received message (because the virtqueue code places completions in
the async queue, and the order in which items are dequeued in
multi-processor machines is not necessarily the same as the order
in which the corresponding closures are invoked).
This change adds sequence numbering to the generic virtqueue code,
and uses this feature in the virtio socket driver to ensure ordered
processing of received packets.