TFS log_write_internal(): fix writing to new log extension
When the current log extension of a TFS filesystem is full and a
new extension is created, any log entries that could not fit in the
current extension must be written to the staging buffer of the new
extension.
This change fixes a regression introduced in commit d151508. In
addition, it fixes a build failure when the TLOG_DEBUG_DUMP
preprocessor #define is present.

Closes #1970.