Aarch64 platform clock: fix conversion to timestamp
When the timer counter value retrieved from the CNTFRQ_EL0 system
register is larger than (1 << 32), the value of `(t << 32) / f`
does not always correspond to the fractional part of the number of
seconds; this is causing incorrect timestamp values to be
retrieved, which on AWS instances can lead to a crash when the ntp
klib is used: "assertion t >= last_raw failed at
/home/circleci/project/src/runtime/clock.h:78 (IP
0xffffffff8002b210)  in now(); halt"
(https://forums.nanovms.com/t/question-on-bootup-messages/32).
This change fixes the timestamp calculation.