ns16550 console driver: fix check of transmission buffer status
The existing code is entering an infinite loop when the
transmission buffer of the ns16550 serial peripheral is non-empty.
This change adds a volatile qualifier to the relevant hardware
register pointer so that the buffer status is retrieved repeatedly
(instead of just once) until it becomes empty. This fixes kernel
stalls that have been observed when running Nanos on AWS Aarch64
instances.