NVMe: fix handling of spurious interrupts on admin queue
On AWS r6g and r7g instances, a spurious interrupt has been
observed to occur on the NVME admin queue during driver
initialization; this was causing the nvme_create_iosq_resp()
closure to deallocate itself when handling the spurious interrupt
(which occurs when the admin completion queue is still empty) and
to be invoked after deallocation when the next interrupt occurs
(i.e. when the queue is no longer empty), leading to an exception
in the interrupt handler and eventually a "frame already full"
crash.
This change refactors the admin interupt handling code so that the
completion queue entry is retrieved by the generic handler, which
invokes the entry-specific handler (with the entry as an argument)
only if the queue is non-empty; in addition, the admin submission
queue head pointer is updated in the generic handler, which removes
duplicated code from the entry-specific handlers.