Aarch64 GIC: add support for operating GICv4.1 controllers.
AWS Graviton3 instances (such as m7g and r7g instances) are
equipped with a GICv4.1 interrupt controller. The GIC driver in the
kernel does not use GICv4.1-specific features, and is able to
operate the controller in the same way as it operates v3.0 and v4.0
controllers.
This change amends the driver initialization code so that it does
not error out when probing a v4.1 controller; this allows the
kernel to run on AWS Graviton3 instances.