Clocks: add support for PTP hardware clock
This change adds a new `clock_now` global variable that can be set
to point to a closure that retrieves the current tiestamp from a
PTP hardware clock.
The NTP klib has been enhanced to be able to retrieve time from the
PTP clock alternatively to using NTP; if the root tuple contains a
"chrony" tuple that has a "refclock" attribute with "ptp" value,
then the NTP klib uses the PTP clock for synchronizing the system
time; if a PTP clock is not available, the klib falls back to NTP.
The next commit will add a PTP clock driver.