id heap: fix alignment of allocated values
The id heap has the property that allocated values are aligned to
the nearest power-of-2 value equal to or greater than the
allocation size. In the current code, it can happen that an
allocation of a given size served by an id heap range whose start
value is not aligned to the allocation size returns a non-aligned
value.
This patch fixes the above issue by removing the padding of the
start argument in the bitmap allocation function (because the start
value is already appropriately aligned by the id heap in the
id_alloc_from_range() function).
In order to optimize performance of bitmap allocations, bitmap
ranges added to an id heap have now a start value always aligned to
BITMAP_WORDLEN (even if the corresponding id heap range is not
aligned), so that the start argument to bitmap_alloc_within_range()
is always aligned to BITMAP_WORDLEN for allocation size values
greater than (BITMAP_WORDLEN / 2). The bitmap_start field of struct
id_range contains the aligned start value of a range, and is used
to apply the appropriate offsets during allocations and
deallocations.
Since now the id heap works correctly with arbitrarily aligned
ranges, we can remove the padding to 2MB of the start value of
physical memory ranges, which was causing a noticeable waste of
memory space in low-memory machines.