Network sockets: fix setting of flags in struct tcp_pcb
The flags field in struct tcp_pcb contains both flags expressing
socket configuration settings (e.g. whether the Nagle algorithm is
enabled), and flags describing the current state of a socket (e.g.
whether the window scale option is active).
The current code in accept_bh() is overwriting both types of flags
in a struct tcp_pcb representing a connection created from a
listening socket; this causes any TCP options activated as part of
the connection establishment handshake to be reset.
This change fixes the above issue by defining a SOCK_TCP_CFG_FLAGS
bitmask that ORs the flags expressing socket configuration
settings, so that only these flags are modified in accept_bh();
this fixes a TCP performance issue that causes the TCP window scale
option to be deactivated after a connection is etablished.