Message strings: remove unneeded "%s" format specifiers
This change amends format strings in various places in order to
remove unneeded "%s" format specifiers, e.g. when they can be
replaced by "%c", or when err_msg() can be called instead of
rprintf(). In addition, "%s" specifiers are being removed in a few
places where there is no corresponding string in the function
argument list.
This is being done in preparation for the removal of unbounded
string functions.