timm(): avoid specifying multiple tuple attributes in one call
This change amends the argument list in all calls to timm() that
specify multiple tuple attributes, so that only the first attibute
is specified in the timm() call, while the additional attributes
are specified via calls to timm_append(). This is done in
preparation for the commit that will remove unbounded string
functions, which will change the timm() implementation to support
only one tuple attribute.