timm(): use `timm_oom` instead of allocating "out of memory" tuples
This change replaces a few timm() calls occurring on out of memory
conditions with references to the `timm_oom` global tuple. This
simplifies the logic in the Unix fault handler to check for out of
memory errors, and removes one use of the (soon to be removed)
buffer_compare_with_cstring_ci() function.