ELF exec: add AT_BASE entry to auxiliary vector
This change adds the AT_BASE entry (which contains the base
address of the interpreter for the user program, or zero if there
is no interpreter) to the auxiliary vector that is set up on the
user stack when executing the program. This is needed by the musl
dynamic linker, which would otherwise calculate a wrong base
address during startup, resulting in a segmentation fault.