E2E tests: ruby: remove sinatra gem install and GEM_HOME env var
The ruby_3.1.2 package already includes the sinatra gem, thus it is
not necessary to install it locally in order to run a Ruby web
server. This change removes the sinatra gem installation step in
the Rubty e2e test; in addition, the GEM_HOME environment variable
is being removed from the Ops configuration file because the Ruby
package already provides a correct value for this variable
(consistent with the contents of the package).
This change fixes a 'Sinatra could not start, the "rackup" gem was
not found' error that occurs when trying to run the (locally
installed) latest release of the sinatra gem during the e2e tests.