vmap_handler: change return type from void to boolean
If a vmap handler returns false, the vmap_iterator() function
terminates immediately.
This change allows a vmap_handler to be cast to a rmnode_handler,
and is made in preparation for the next commit that will remove
code duplication by refactoring the code that walks a process
virtual memory map.