mmap(): add vmap_range_walk() helper function
This new function walks a given range in the process memory map and
invokes a vmap handler for each mapping encountered in that range.
If the `allow_gaps` argument is false, the presence of any gap in
the memory map over the given range makes this function return
RM_ABORT; otherwise, this function returns RM_MATCH if at least
(part of) one mapping exists in the given range, RM_NOMATCH of no
mapping exists.
The existing code in src/unix/mmap.c is refactored to use the new
function and remove code duplication.