vmap_remove_intersection(): fix node offset passed to vmap_handler
When a memory mapping is shrunk so that only its tail is left, the
node offset of the relevant vmap struct should be advanced to
account for the removed range; however, if the modification of the
node offset is done before calling altered_vmap_key(), the vmap
passed to the vmap_handler has a bogus node offset. This can cause
unmapping of file-backed address ranges to deallocate wrong
physical pages.
This change fixes the above issue by moving the update of the node
offset to after the invocation of the vmap_handler.