Makefiles: move Make variable initialization out of rules.mk
This change moves the initialization of global Make variables from
rules.mk to a new vars.mk file. This allows specifying that a klib
or a test should be built or run only in certain environments or
configurations (for example, only for a given architecture).
The RUNTIME_TESTS target in the main Makefile has been modified by
adding the missing symlink program and making the vsyscall program
conditional to the x86_64 architecture.