Hyper-V: netvsc: fix setting of pbuf length in received packets
The existing code is setting the tot_len field of the first struct
of a pbuf chain to a value double the correct value, and is failing
to update the tot_len value of all pbufs in a chain when appending
received data.
This change fixes the above issues, which were causing a page fault
in the pbuf_realloc() lwIP function during reception of an IPv6
packet.