aio: io_submit(): fix reference counting of process context
A process context retrieved via a get_process_context() call has an
initial refcount value of 1; if the refcount is not decremented in
the relevant completion closure, the context will never be returned
to the free list, nor deallocated.
This change adds a context field to the aio_complete and
aio_eventfd_complete closures, so that the process context can be
properly released.