io_uring_enter(): fix reference counting of process context
A process context retrieved via a get_process_context() call has an
initial refcount value of 1; if the refcount is not decremented
when the async operations triggered by io_uring_enter() are
completed, the context will never be returned to the free list, nor
deallocated.
This change adds a context field to the iour_rw_complete and
iour_close_complete closures, so that the process context can be
properly released.