virtIO: implement optional notification of queue status change
This change adds a `kick` parameter to the vqmsg_commit_seqno()
function; this parameter specifies whether a virtIO device should
be notified of the new message on the virtqueue. The virtio_net
and virtio_socket drivers have been modified so that their
virtqueues can be refilled with multiple messages at a time (until
there are no free entries), and the `kick` argument is set to true
only for the last message; this allows refilling completely a
virtqueue after a temporary failure to push new messages (e.g.
because of a memory allocation failure), and avoids unnecessary
notifications to the virtIO device, which are costly because they
usually involve a VM exit.