virtio_net: adapt RX buffer length to device features
When TCP segmentation (or UDP fragmentation) offload features are
enabled, a received packet can be composed of multiple TCP (or UDP)
packets, and thus exceed the size of RX buffers as they are
allocated in the existing code; if RX buffers are not mergeable,
each single buffer must be large enough to hold a received packet,
otherwise the device is unable to copy received packets to the RX
virtqueue.
This change modifies the virtio_net driver initialization code so
that the length of RX buffers is increased (to the maximum value
supported by lwIP) when large buffers are required. This fixes
"Receiving buffer is too small to hold frame of current size" error
messages that have been observed when running Nanos under AWS
Firecracker.

Closes #1999.