Security: add kernel address space layout randomization (KASLR)
This change set adds a security feature that randomizes the
virtual address ranges where the kernel and klib binaries are
mapped. The range from which random addresses are generated spans
2GB, and the only restriction on generated addresses is 4KB
alignment, therefore the number of possible addresses is 2^19.

The kernel binary is now created as a PIE (position-independent
executable), and during boot the kernel applies dynamic relocations
to its code when moving to a random address. The dynamic section
is specified in the linker scripts as part of the read-only data
section (thereby preventing the default linker behavior that places
the dynamic section in the read-write data section), so that it's
mapped with read-only permissions.
Platform-specific code that maps the kernel to its final virtual
address has been removed, since the final virtual address is now
generated when doing KASLR.
The initial mappings set up during early boot by the hypervisor or
bootloader or kernel platform-specific code are removed in the
reclaim_regions() function (or earlier in the boot process).
In the x86_64 code, the assembly initialization code has been moved
to a new init.s file, and is now bound to the .start linker section
so that its load address is known at compile time; this prevents
the assembler from creating static relocation entries that cannot
be used when making a PIE. The initial pages area is now identity-
mapped instead of using the PAGES_BASE virtual address; this allows
removing the `bootstrapping` global variable from the generic page
table code.
In the aarch64 linker script, the vvar section has been moved after
the data sections, so that it's correctly mapped with read-write
permissions when doing KASLR.
The RISC-V assembly initialization code has been moved to a new
init.S file, which is compiled as non-position-independent code, so
that it can use the build-time address of the dynamic linker
section to apply the initial relocations that are needed to access
global variables. In the RISC-V linker script, sections have been
moved around in order to ensure that read-write data is located
between the READONLY_END and bss_start linker symbols.