filesystem_creat_unnamed(): move to generic filesystem code
This change allows creating temporary files (i.e. files that are
not associated to a directory entry and are deleted when closed) on
filesystems other than TFS. The create callback of existing
filesystem drivers has been amended to be able to create temporary
files (or return FS_STATUS_INVAL if this feature is not supported).
This change will be needed when support for memfd_create() is
added.