Syscall initialization: add init_syscall() macro
This macro is a more lightweight version of register_syscall(), and
can be used to initialize the name and flags of a syscall without
configuring a handler. It is now used for all non-supported
syscalls.
If a klib implementing additional syscalls is loaded, it can call
swap_syscall_handler() instead of register_syscall().