uname: add configurability of sysname and release
With this change, it is possible to configure the kernel to insert
arbitrary strings in the `sysname` and `release` fields of the
struct used in the uname syscall. This allows running user programs
that expect these fields to have specific values (e.g. programs
that expect the sysname to be "Linux", or programs that parse the
release string to infer a Linux kernel version).
To configure the sysname and release fields, a "uname"
configuration tuple is used: this tuple can contain a "sysname"
string and a "release" string, which allow configuring the
respective fields; if a given string is not present in the
configuration, its default value is used (i.e., "Nanos" for the
sysname field, and "5.0-<NANOS_VERSION>" for the release field).
Example snippet of Ops configuration file:
```
"ManifestPassthrough": {
  "uname": {
    "sysname": "Linux",
    "release": "5.0.5"
  }
}
```