kernel init: ensure stage3 startup is done after probing devices
With this change, the stage3 startup() function is always invoked
after all attached peripherals have been probed, even if the root
filesystem is loaded before PCI bus discovery is complete (which
can happen if disk device interrupts are handled on a CPU other
than the boot CPU).
The next commit will allow device interrupts to be assigned to an
arbitrary CPU, and this commit prevents errors such as "NET: no
network interface found" from occurring if network device probing
is not complete before the root filesystem is loaded.