Interrupts: implement CPU affinity
In the current code, all device interrupts are routed to either
the boot CPU (on x86 and ARM), or to any CPU (on RISC-V).
This new feature allows interrupts to be configured to target a
specific CPU, which allows optimizing interrupt allocation to
spread a workload over multiple CPUS.
The arch-specific msi_format() and dev_irq_enable() functions have
been changed to take an additional `target_cpu` parameter, while
msi_get_vector() has been replaced by msi_get_config() which allows
retrieving both the interrupt number and the target CPU associated
to a given PCI MSI. Other changes in arch-specific interrupt code
have been made in order to support targeting an arbitrary interrupt
to an arbitrary CPU.
Various functions in the PCI and virtIO code have been amended to
take an additional `cpu_affinity` range parameter, which is used as
a hint to select a target CPU when a given interrupt vector is
enabled in the interrupt controller. The irq_get_target_cpu()
utility function takes a CPU affinity range as parameter and
returns an optimal target CPU, selected from the supplied range
based on which CPUs have the smallest number of device interrupts
targeted at them, so as to spread the interrupt handling work over
different CPUs. The irq_put_target_cpu() function is called to
signal that an interrupt has been removed from a given target CPU.