virtio-net: add multiqueue support
This change enhances the virtio-net driver to support multiple
tx/rx queues. This allows achieving better network performance in
multi-CPU instances running workloads that serve multiple network
connections simultaneously.
By default, the virtio-net driver uses as many queues as supported
by the attached device; it is possible to override this behavior by
specifying the "io-queues" configuration option in the manifest
tuple corresponding to a given network interface. For example,
the following snippet of an Ops configuration file instructs the
driver to use 2 queues for the first network interface:
```
"en1": {
  "io-queues": "2"
}
```
The number of queues used by the driver is always limited to the
number of CPUs in the running instance (this behavior cannot be
overridden by the "io-queues" option).
In order to optimize parallelism, each tx/rx queue is configured
with an interrupt affinity such that different queues are served
by different CPUs.